+++ earth2studio/models/dx/identity.py
@@ -19,7 +19,7 @@ from collections import OrderedDict
 import numpy as np
 import torch
 
-from earth2studio.models.batch import batch_func
+from earth2studio.models.batch import batch_coords, batch_func
 from earth2studio.utils.type import CoordSystem
 
 
@@ -44,17 +44,23 @@ class Identity(torch.nn.Module):
         """
         return OrderedDict({"batch": np.empty(0)})
 
-    @property
-    def output_coords(self) -> CoordSystem:
-        """Ouput coordinate system of diagnostic model, time dimension should contain
-        time-delta objects
+    @batch_coords()
+    def output_coords(self, input_coords: CoordSystem) -> CoordSystem:
+        """Output coordinate system of diagnostic model
+
+        Parameters
+        ----------
+        input_coords : CoordSystem
+            Input coordinate system to transform into output_coords
+            by default None, will use self.input_coords.
 
         Returns
         -------
         CoordSystem
             Coordinate system dictionary
         """
-        return OrderedDict({"batch": np.empty(0)})
+
+        return input_coords.copy()
 
     @torch.inference_mode()
     @batch_func()
@@ -64,5 +70,4 @@ class Identity(torch.nn.Module):
         coords: CoordSystem,
     ) -> tuple[torch.Tensor, CoordSystem]:
         """Forward pass of diagnostic"""
-        output_coords = coords.copy()
-        return x, output_coords
+        return x, self.output_coords(coords)