+++ earth2studio/perturbation/zero.py
@@ -43,6 +43,6 @@ class Zero:
         Returns
         -------
         tuple[torch.Tensor, CoordSystem]:
-            Ouput tensor and respective coordinate system dictionary
+            Output tensor and respective coordinate system dictionary
         """
         return x, coords