+++ test/models/px/test_dlwp.py
@@ -103,9 +103,9 @@ def test_dlwp_call(time, dlwp_phoo_cs_transform, device):
         time = [time]
 
     assert out.shape == torch.Size(
-        [len(time), 1, len(p.output_coords["variable"]), 721, 1440]
+        [len(time), 1, len(p.output_coords(p.input_coords)["variable"]), 721, 1440]
     )
-    assert (out_coords["variable"] == p.output_coords["variable"]).all()
+    assert (out_coords["variable"] == p.output_coords(coords)["variable"]).all()
     assert (out_coords["time"] == time).all()
     assert torch.allclose(
         out, p.to_equirectangular(p.to_cubedsphere(x[:, 1:] + 6))
@@ -174,9 +174,11 @@ def test_dlwp_iter(ensemble, dlwp_phoo_cs_transform, device):
     for i, (out, out_coords) in enumerate(p_iter):
         assert len(out.shape) == 6
         assert out.shape[0] == ensemble
-        assert (out_coords["variable"] == p.output_coords["variable"]).all()
+        assert (
+            out_coords["variable"] == p.output_coords(p.input_coords)["variable"]
+        ).all()
         assert (out_coords["time"] == time).all()
-        assert out_coords["lead_time"][0] == np.timedelta64(6 * (i + 1), "h")
+        assert out_coords["lead_time"] == np.timedelta64(6 * (i + 1), "h")
         assert torch.allclose(
             out, p.to_equirectangular(p.to_cubedsphere(x[:, :, 1:] + 6 * (i + 1)))
         )  # Need to cs transform here to get right values
@@ -266,7 +268,7 @@ def test_dlwp_package(device, model_cache_context):
         time = [time]
 
     assert out.shape == torch.Size([len(time), 1, 7, 721, 1440])
-    assert (out_coords["variable"] == p.output_coords["variable"]).all()
+    assert (out_coords["variable"] == p.output_coords(coords)["variable"]).all()
     assert (out_coords["time"] == time).all()
     handshake_dim(out_coords, "lon", 4)
     handshake_dim(out_coords, "lat", 3)